% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_curve.R
\name{as.data.frame.std_calc}
\alias{as.data.frame.std_calc}
\title{Convert \code{std_calc} to data frame}
\usage{
\method{as.data.frame}{std_calc}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{std_calc}, the output of \code{std_curve_calc()}}

\item{row.names}{Optional vector of rownames.}

\item{optional}{logical. If TRUE, setting row names and converting column
names (to syntactic names: see make.names) is optional. Note that all of
R's base package as.data.frame() methods use optional only for column names
treatment, basically with the meaning of data.frame(*, check.names =
!optional). See also the make.names argument of the matrix method.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
data.frame
}
\description{
Convert \code{std_calc} to data frame
}
