% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_standard_curve.R
\name{as.data.frame.std_prediction}
\alias{as.data.frame.std_prediction}
\title{Title}
\usage{
\method{as.data.frame}{std_prediction}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{optional}{}
}
\value{

}
\description{
Title
}
