% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_curve.R
\name{calc_std_curve}
\alias{calc_std_curve}
\title{Generate Points for Plotting a Standard Curve}
\usage{
calc_std_curve(std_curve)
}
\arguments{
\item{std_curve}{A model generated from \code{std_curve_fit()}}
}
\value{

}
\description{
Generate Points for Plotting a Standard Curve
}
