% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_standard_curve.R
\name{fit_standard_curve}
\alias{fit_standard_curve}
\title{Create a Standard Curve From Known Data}
\usage{
fit_standard_curve(data, conc, resp)
}
\arguments{
\item{data}{A \code{data.frame} that contains the columns for concentration and observed response for the standard curve.}

\item{conc}{Name of the column that contains the concentration for the standard curve.}

\item{resp}{Name of the column that contains the response values for the standard curve.}
}
\value{
A linear model object to be used as a standard curve, for use with
\code{standard::predict_from_curve()} \code{broom::augment()} or \code{stats::predict()}.
}
\description{
Create a Standard Curve From Known Data
}
