% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_decimal}
\alias{n_decimal}
\title{Finds Minimum Number of Decimal Places}
\usage{
n_decimal(x)
}
\arguments{
\item{x}{Number to calculate decimal places.}
}
\description{
Finds Minimum Number of Decimal Places
}
