% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_curve.R
\name{plot.std_calc}
\alias{plot.std_calc}
\title{Generic Function for Plotting Standard Curve Calculations}
\usage{
\method{plot}{std_calc}(x, ...)
}
\arguments{
\item{x}{output of \code{std_curve_calc()}}

\item{...}{Additional arguments to be passed to \code{std_curve_plot()}}
}
\value{
ggplot2 plot
}
\description{
Generic Function for Plotting Standard Curve Calculations
}
