% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_curve.R
\name{plot.std_curve}
\alias{plot.std_curve}
\title{Generic Function for Plotting Fitted Standard Curves}
\usage{
\method{plot}{std_curve}(x, ...)
}
\arguments{
\item{x}{output of \code{std_surve_fit()}}

\item{...}{Additional arguments to be passed to \code{std_curve_plot()}}
}
\value{
ggplot2 plot
}
\description{
Generic Function for Plotting Fitted Standard Curves
}
