% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_standard_curve.R
\name{predict_from_curve}
\alias{predict_from_curve}
\title{Use a Standard Curve to Calculate Unkown Values}
\usage{
predict_from_curve(model, unkowns)
}
\arguments{
\item{model}{A linear model, created with either \code{lm()} or
\code{standard::fit_standard_curve()}}

\item{unkowns}{A numeric vector of unkown values, which the standard curve
will be used to predict their values.}
}
\value{
a \link[tibble:tibble-package]{tibble} with a column for the unkown
values, and a column \code{.fitted} for the predicted values, based on the
standard curve.
}
\description{
Use a Standard Curve to Calculate Unkown Values
}
