% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_curve.R
\name{print.std_calc}
\alias{print.std_calc}
\title{Printing Results of \code{std_curve_calc()}}
\usage{
\method{print}{std_calc}(x, ...)
}
\arguments{
\item{x}{object of class \code{std_calc}, the output of \code{std_curve_calc}}

\item{...}{additional arguments to be passed to or from methods.}
}
\description{
Printing Results of \code{std_curve_calc()}
}
