% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_standard_curve.R
\name{std_curve_plot}
\alias{std_curve_plot}
\title{Plot a Standard Curve}
\usage{
std_curve_plot(data)
}
\arguments{
\item{data}{Result of \code{std_curve_pred()} or \code{std_curve_fit()}.}
}
\value{
a \link[ggplot2:ggplot]{ggplot2} plot with the standard curve and unkowns
plotted, whch can be further customised using \code{ggplot} options.
}
\description{
Plot a Standard Curve
}
