% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_curve.R
\name{[.std_calc}
\alias{[.std_calc}
\title{Generic function for subsetting output of \code{std_curve_fit()}}
\usage{
\method{[}{std_calc}(x, i, j)
}
\arguments{
\item{x}{object of class \code{std_curve}, the output of \code{std_curve_fit()}}

\item{i}{row index}

\item{j}{column index}
}
\value{
column of tibble
}
\description{
Generic function for subsetting output of \code{std_curve_fit()}
}
