% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_curve.R
\name{std_curve_fit}
\alias{std_curve_fit}
\title{Create a Standard Curve From Known Data}
\usage{
std_curve_fit(data, conc, resp, curve = "linear")
}
\arguments{
\item{data}{A \code{data.frame} that contains the columns for concentration and
observed response for the standard curve.}

\item{conc}{Name of the column that contains the concentration for the
standard curve.}

\item{resp}{Name of the column that contains the response values for the
standard curve.}

\item{curve}{}
}
\value{
A linear model ( \link[stats:lm]{lm()} ) object to be used as a standard
curve, for use with \code{standard::std_curve_calc()} \code{broom::augment()} or
\code{stats::predict()}.
}
\description{
Create a Standard Curve From Known Data
}
\examples{
library(standard)

# Protein concentrations of the standards used in the assay
prot <- c(
  0.000, 0.016, 0.031, 0.063, 0.125, 0.250, 0.500, 1.000,
  0.000, 0.016, 0.031, 0.063, 0.125, 0.250, 0.500, 1.000
)

# absorbance readins from the standards used in the assay
abs <- c(
  0.329, 0.352, 0.349, 0.379, 0.417, 0.491, 0.668, 0.956,
  0.327, 0.341, 0.355, 0.383, 0.417, 0.446, 0.655, 0.905
)
assay_data <- data.frame(
  Protein = prot,
  Absorbance = abs
)

# unknown concentrations
unk <- c(0.554, 0.568, 0.705)


assay_data \%>\%
  std_curve_fit(Protein, Absorbance) \%>\%
  plot()
}
