% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_standard_curve.R
\name{std_curve_plot}
\alias{std_curve_plot}
\title{Plot a Standard Curve}
\usage{
std_curve_plot(data)
}
\arguments{
\item{data}{Result of \code{std_curve_pred()} or \code{std_curve_fit()}.}
}
\value{
a \link[ggplot2:ggplot]{ggplot2} plot with the standard curve and unkowns
plotted, whch can be further customised using \code{ggplot} options.
}
\description{
Plot a Standard Curve
}
\examples{
library(standard)

# Protein concentrations of the standards used in the assay
prot <- c(0.000, 0.016, 0.031, 0.063, 0.125, 0.250, 0.500, 1.000,
          0.000, 0.016, 0.031, 0.063, 0.125, 0.250, 0.500, 1.000)

# absorbance readins from the standards used in the assay
abs <- c(0.329, 0.352, 0.349, 0.379, 0.417, 0.491, 0.668, 0.956,
         0.327, 0.341, 0.355, 0.383, 0.417, 0.446, 0.655, 0.905)
assay_data <- data.frame(
  Protein = prot,
  Absorbance = abs
)

# unknwon concentrations
unk <- c(0.554, 0.568, 0.705)


assay_data \%>\%
  std_curve_fit(Protein, Absorbance) \%>\%
  std_curve_calc(unk) \%>\%
  plot()
}
