% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_standard_curve.R
\name{std_paste_formula}
\alias{std_paste_formula}
\title{Extract and Paste Formula From Standard Curve}
\usage{
std_paste_formula(std_curve, digits = 3)
}
\arguments{
\item{digits}{}
}
\value{
a string of the extracted formular from the standard curve
}
\description{
Extract and Paste Formula From Standard Curve
}
